Streamline Confluence space import
