+++ .gitignore
@@ -1,7 +1,8 @@
 /.idea/
 /.env
-.DS_Store
-*.pyc
-*.ipynb
-**/__pycache__/
+__pycache__
 .coverage
+.DS_Store
+*.ipynb
+*.log
+*.pyc
