+++ configuration.py
@@ -22,7 +22,6 @@ logging.log(logging.DEBUG, f"Project path: {project_path}")
 
 chart_folder_path = os.path.join(project_path, "content", "charts")
 sql_file_path = os.path.join(project_path, "content", "database", "confluence_pages_sql.db")
-persist_page_processing_queue_path = os.path.join(project_path, "content", "queues", "confluence_page_processing_queue")
 vector_folder_path = os.path.join(project_path, "content", "vectors", "confluence_pages")
 interactions_folder_path = os.path.join(project_path, "content", "vectors", "confluence_interactions")
