+++ open_ai/assistants/openai_assistant.py
@@ -200,7 +200,3 @@ def user_interaction(client):
 def load_manage_assistants():
     client = initiate_client()
     user_interaction(client)
-
-
-if __name__ == "__main__":
-    load_manage_assistants()
