+++ pyproject.toml
@@ -8,7 +8,6 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.12"
-persist-queue = "^0.8.1"
 openai = "^1.12.0"
 langchain = "^0.1.7"
 chromadb = "^0.4.22"
