--- shams_gpt/code_index.txt
@@ -1,107 +1,0 @@
-Below is a comprehensive and detailed list of the documents related to the 'Nur' project, including all their capabilities and dependencies, without resorting to generalized 'etc.' descriptions. Here's the exhaustive list:
-
-main.py
-
-Capabilities:
-Orchestrates overall functionality, integrating various modules.
-Manages Confluence integration, vector operations, query handling, and Slack bot interactions.
-Acts as the primary entry point, coordinating the workflow between different components.
-Dependencies:
-Internal Modules: confluence_integration.retrieve_space, vector.chroma_threads, oai_assistants.query_assistant_from_documents, gpt_4t.query_from_documents_threads, slack.channel_interaction_threads, confluence_integration.extract_page_content_and_store_processor.
-vectorize_and_persist_processor.py
-
-Capabilities:
-Processes and vectorizes data for efficient storage and retrieval.
-Implements algorithms for data transformation and vectorization.
-Dependencies:
-Python Libraries: numpy for numerical operations, sqlalchemy for database interactions.
-extract_page_content_and_store_processor.py
-
-Capabilities:
-Extracts and processes content from web pages, likely for data ingestion or web scraping.
-Uses parsing techniques to extract meaningful information from HTML content.
-Dependencies:
-Libraries: beautifulsoup4 for HTML parsing, requests for HTTP requests.
-confluence_client.py
-
-Capabilities:
-Handles API interactions with Confluence, including data retrieval and posting.
-Implements methods for accessing and manipulating Confluence content.
-Dependencies:
-Libraries: requests for HTTP communications, beautifulsoup4 for parsing HTML responses.
-requirements.txt
-
-Capabilities:
-Lists all Python dependencies required for the project.
-Ensures consistent development and deployment environments.
-Dependencies:
-Libraries: openai, requests, schedule, sqlalchemy, beautifulsoup4, chromadb, langchain, tiktoken, prettytable, slack_sdk, numpy, deprecated.
-channel_interaction_threads.py
-
-Capabilities:
-Manages threaded interactions within communication channels, primarily focusing on Slack integration.
-Implements concurrent handling of user interactions or bot responses.
-Dependencies:
-Libraries: slack_sdk for Slack API interactions, Python's threading capabilities for concurrency.
-event_publisher.py
-
-Capabilities:
-Publishes events across the system, possibly for inter-module communication or external notifications.
-Implements an event-driven architecture to trigger actions based on events.
-Dependencies:
-Event handling and messaging related libraries, potentially including event queues and asynchronous communication tools.
-nur_database.py
-
-Capabilities:
-Manages all database operations, including CRUD operations and database connection management.
-Acts as a central point for data persistence and retrieval.
-Dependencies:
-Database handling libraries, primarily sqlalchemy.
-sync_up_qa_articles_to_confluence.py
-
-Capabilities:
-Syncs QA articles with Confluence, potentially for documentation updates or knowledge sharing.
-Implements routines to ensure data consistency between local content and Confluence pages.
-Dependencies:
-Libraries: requests for HTTP API interactions, possibly Confluence API-specific libraries.
-retrieve_space.py
-
-Capabilities:
-Retrieves and processes data from specified spaces in Confluence.
-Used for bulk data extraction or specific data queries.
-Dependencies:
-Libraries: beautifulsoup4 for parsing, requests for API interactions.
-event_consumer_threads.py
-
-Capabilities:
-Manages event consumption in a multi-threaded environment.
-Handles incoming events, possibly from a message queue or an external trigger.
-Dependencies:
-Libraries and tools for handling threading and event processing.
-query_from_documents_threads.py
-
-Capabilities:
-Handles document-based queries using threading, potentially for search operations or data retrieval.
-Implements concurrent querying to enhance performance.
-Dependencies:
-Document processing and multi-threading related libraries.
-dynamic_executor.py
-
-Capabilities:
-Dynamically executes tasks based on predefined triggers or conditions.
-Flexible execution model to respond to various operational scenarios.
-Dependencies:
-Task scheduling and execution libraries, possibly including schedule or similar tools.
-chroma_threads.py
-
-Capabilities:
-Manages threads for operations involving 'Chroma', which might be related to advanced data processing, analytics, or visualization.
-Ensures efficient handling of resource-intensive tasks.
-Dependencies:
-Libraries: chromadb for data processing, Python threading tools.
-README.md
-
-Capabilities:
-Provides comprehensive documentation and an overview of the project.
-Includes setup instructions, usage guidelines, and contribution details.
-This list represents a thorough indexing of each component within the 'Nur' project, detailing their specific roles, functionalities, and the technologies they depend upon. It serves as a detailed reference for understanding the project's architecture and for navigating its various parts.
